Fix random.real call in Strand, and rename genetics/ to neuroevolution/
