+++ src/modules/neuroevolution/ConnectionGene.js
