+++ src/modules/neuroevolution/NodeGene.js
