+++ src/modules/neuroevolution/Sequencer.js
