+++ src/modules/neuroevolution/connectionGene.test.js
